moved the Configuration class to a new package before adding a DB specific configuration class for the control utility
