+++ src/main/java/com/cisco/josouthe/ETLTransferMain.java
@@ -1,5 +1,6 @@
 package com.cisco.josouthe;
 
+import com.cisco.josouthe.config.Configuration;
 import com.cisco.josouthe.scheduler.MainControlScheduler;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
