+++ src/main/java/com/cisco/josouthe/scheduler/DatabaseInsertTask.java
@@ -1,6 +1,6 @@
 package com.cisco.josouthe.scheduler;
 
-import com.cisco.josouthe.Configuration;
+import com.cisco.josouthe.config.Configuration;
 import com.cisco.josouthe.database.Database;
 import com.cisco.josouthe.exceptions.FailedDataLoadException;
 import com.cisco.josouthe.util.WorkingStatusThread;
